backend/authentication_service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
asyncpg
authlib
python-dotenv

